+++ android/src/main/java/com/plugin/tim_push/activity/TimPushActivity.java
@@ -1,12 +1,12 @@
-package com.example.tim_push.activity;
+package com.plugin.tim_push.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 
-import com.example.tim_push.application.TimPushApplication;
-import com.example.tim_push.common.Extras;
+import com.plugin.tim_push.application.TimPushBootstrap;
+import com.plugin.tim_push.common.Extras;
 
 import io.flutter.embedding.android.FlutterActivity;
 
@@ -15,7 +15,7 @@ public class TimPushActivity extends FlutterActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         handleIntent(getIntent());
-        TimPushApplication.hadLaunchedMainActivity = true;
+        TimPushBootstrap.markMainActivityLaunched();
     }
 
     @Override
@@ -33,13 +33,6 @@ public class TimPushActivity extends FlutterActivity {
 
     @Override
     public String getCachedEngineId() {
-        if (TimPushApplication.useCustomFlutterEngine) {
-            return Extras.FLUTTER_ENGINE;
-        }
-        try {
-            return getIntent().getStringExtra("cached_engine_id");
-        } catch (Exception ignored) {
-            return null;
-        }
+        return TimPushBootstrap.getCachedEngineId(getIntent());
     }
 }